src/ingest/requirements-local.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
boto3
python-dotenv